feat: migrate ReportAvatar screen to AttachmentModalScreen
