+++ src/libs/Navigation/AppNavigator/AuthScreens.tsx
@@ -86,7 +86,6 @@ const loadTrackExpensePage = () => require<ReactComponentModule>('../../../pages
 const loadSubmitExpensePage = () => require<ReactComponentModule>('../../../pages/SubmitExpensePage').default;
 const loadProfileAvatar = () => require<ReactComponentModule>('../../../pages/settings/Profile/Avatar/ProfileAvatar').default;
 const loadWorkspaceAvatar = () => require<ReactComponentModule>('../../../pages/workspace/WorkspaceAvatar').default;
-const loadReportAvatar = () => require<ReactComponentModule>('../../../pages/ReportAvatar').default;
 const loadReceiptView = () => require<ReactComponentModule>('../../../pages/TransactionReceiptPage').default;
 const loadWorkspaceJoinUser = () => require<ReactComponentModule>('@pages/workspace/WorkspaceJoinUserPage').default;
 
@@ -582,6 +581,12 @@ function AuthScreens() {
                     listeners={modalScreenListeners}
                 />
                 <RootStack.Screen
+                    name={SCREENS.REPORT_AVATAR}
+                    options={attachmentModalScreenOptions}
+                    getComponent={loadAttachmentModalScreen}
+                    listeners={modalScreenListeners}
+                />
+                <RootStack.Screen
                     name={SCREENS.PROFILE_AVATAR}
                     options={{
                         headerShown: false,
@@ -598,15 +603,6 @@ function AuthScreens() {
                         presentation: Presentation.TRANSPARENT_MODAL,
                     }}
                     getComponent={loadWorkspaceAvatar}
-                    listeners={modalScreenListeners}
-                />
-                <RootStack.Screen
-                    name={SCREENS.REPORT_AVATAR}
-                    options={{
-                        headerShown: false,
-                        presentation: Presentation.TRANSPARENT_MODAL,
-                    }}
-                    getComponent={loadReportAvatar}
                     listeners={modalScreenListeners}
                 />
                 <RootStack.Screen
