+++ src/pages/media/AttachmentModalScreen/routes/report/ReportAvatarModalContent.tsx
@@ -1,19 +1,17 @@
 import React, {useMemo} from 'react';
-import AttachmentModal from '@components/AttachmentModal';
 import useOnyx from '@hooks/useOnyx';
-import Navigation from '@libs/Navigation/Navigation';
-import type {PlatformStackScreenProps} from '@libs/Navigation/PlatformStackNavigation/types';
-import type {AuthScreensParamList} from '@libs/Navigation/types';
 import {getDefaultGroupAvatar, getPolicyName, getReportName, getWorkspaceIcon, isGroupChat, isThread, isUserCreatedPolicyRoom} from '@libs/ReportUtils';
 import {getFullSizeAvatar} from '@libs/UserUtils';
+import type {AttachmentModalBaseContentProps} from '@pages/media/AttachmentModalScreen/AttachmentModalBaseContent/types';
+import AttachmentModalContainer from '@pages/media/AttachmentModalScreen/AttachmentModalContainer';
+import useDownloadAttachment from '@pages/media/AttachmentModalScreen/routes/hooks/useDownloadAttachment';
+import type {AttachmentModalScreenProps} from '@pages/media/AttachmentModalScreen/types';
 import ONYXKEYS from '@src/ONYXKEYS';
-import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
 
-type ReportAvatarProps = PlatformStackScreenProps<AuthScreensParamList, typeof SCREENS.REPORT_AVATAR>;
-
-function ReportAvatar({route}: ReportAvatarProps) {
+function ReportAvatarModalContent({navigation, route}: AttachmentModalScreenProps<typeof SCREENS.REPORT_AVATAR>) {
     const {reportID, policyID} = route.params;
+
     const [report] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${reportID}`, {canBeMissing: false});
     const [policy] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY}${policyID}`, {canBeMissing: true});
     const [isLoadingApp = true] = useOnyx(ONYXKEYS.IS_LOADING_APP, {canBeMissing: true});
@@ -26,6 +24,7 @@ function ReportAvatar({route}: ReportAvatarProps) {
                 isWorkspaceAvatar: false,
             };
         }
+
         if (isUserCreatedPolicyRoom(report)) {
             return {
                 source: report?.avatarUrl ? getFullSizeAvatar(report.avatarUrl, 0) : undefined,
@@ -35,30 +34,33 @@ function ReportAvatar({route}: ReportAvatarProps) {
         }
 
         return {
-            source: getFullSizeAvatar(getWorkspaceIcon(report, policy).source, 0),
+            source: getFullSizeAvatar(getWorkspaceIcon(report).source, 0),
             headerTitle: getPolicyName({report, policy}),
             // In the case of default workspace avatar, originalFileName prop takes policyID as value to get the color of the avatar
             originalFileName: policy?.originalFileName ?? policy?.id ?? report?.policyID,
             isWorkspaceAvatar: true,
         };
-    }, [report, policy]);
+    }, [policy, report]);
+
+    const onDownloadAttachment = useDownloadAttachment();
+
+    const contentProps = useMemo<AttachmentModalBaseContentProps>(
+        () => ({
+            ...attachment,
+            shouldShowNotFoundPage: !report?.reportID && !isLoadingApp,
+            isLoading: (!report?.reportID || !policy?.id) && !!isLoadingApp,
+            onDownloadAttachment,
+        }),
+        [attachment, isLoadingApp, policy?.id, report?.reportID, onDownloadAttachment],
+    );
 
     return (
-        <AttachmentModal
-            headerTitle={attachment.headerTitle}
-            defaultOpen
-            source={attachment.source}
-            onModalClose={() => {
-                Navigation.goBack(report?.reportID ? ROUTES.REPORT_WITH_ID_DETAILS.getRoute(report?.reportID) : undefined);
-            }}
-            isWorkspaceAvatar={attachment.isWorkspaceAvatar}
-            maybeIcon
-            originalFileName={attachment.originalFileName}
-            shouldShowNotFoundPage={!report?.reportID && !isLoadingApp}
-            isLoading={(!report?.reportID || !policy?.id) && !!isLoadingApp}
+        <AttachmentModalContainer
+            navigation={navigation}
+            contentProps={contentProps}
         />
     );
 }
+ReportAvatarModalContent.displayName = 'ReportAvatarModalContent';
 
-ReportAvatar.displayName = 'ReportAvatar';
-export default ReportAvatar;
+export default ReportAvatarModalContent;
